Create Web and mobile Recipe Finder
Food is one of the basic needs of humans. In today's generation, cooking has become a lot easier given that there are a lot of recipes available. A recipe basically gives instructions on how to prepare and cook food. The proposed project will be a platform where different food recipes are displayed and stored. The main goal of this project is to implement a web and mobile based food recipe finder. The project aims at the automation of finding food recipes. The system will ease up and simplify the process which will save valuable time and effort for users.